Simplify polling
Previously I was trying to poll in two stages: first to grab post ids
and then to grab the post data on demand. This didn't seem to be as
useful as I had anticipated for two reasons: we're still limited to
grabbing post ids in batches of 100 due to the WordPress.com API; we
don't know in advance if a post has attachments so a bunch of stuff with
the fetching and paging logic was going to be messed up anyway.

Now with fewer lines of code and less complexity this is easier to work
with.