Add infinite srolling and network indicator
This was a big jump and involved adding a new JavaScript port. I've made
a change here to support the port in development with the reactor
without adding any further build steps: we need to load the
`reactor.html` from the reactor's dashboard instead of `src/Main.elm`.
This is not a big change but it does mean we lose the fullscreen error
displays in the browser when we mess up. If the page doesn't load we'll
need to open `src/Main.elm` in another browser tab via the reactor.